requirements.txt
Flask==2.2.5
python-dotenv~=1.0.1
graphene~=2.1.9
pandas~=2.2.3
graphql-core~=2.3.2
psycopg2~=2.9.10
numpy~=2.0.2
tensorflow~=2.18.0
tf_keras~=2.18.0
matplotlib~=3.9.2
seaborn~=0.13.2